requirements.txt
gradio
numpy
pillow
transformers
torch
